requirements.txt
prettytable
colorama
dnspython
requests
